+++ test/doing_sections_test.rb
@@ -1,0 +1,49 @@
+require 'fileutils'
+require 'tempfile'
+require 'time'
+require 'yaml'
+
+require 'helpers/doing-helpers'
+require 'test_helper'
+
+# Tests for entry modifying commands
+class DoingSectionTest < Test::Unit::TestCase
+  include DoingHelpers
+  ENTRY_REGEX = /^\d{4}-\d\d-\d\d \d\d:\d\d \|/.freeze
+
+  def setup
+    @tmpdirs = []
+    @result = ''
+    @basedir = mktmpdir
+    @wwid_file = File.join(@basedir, 'wwid.md')
+    @backup_dir = File.join(@basedir, 'doing_backup')
+    @config_file = File.join(File.dirname(__FILE__), 'test.doingrc')
+  end
+
+  def teardown
+    FileUtils.rm_rf(@tmpdirs)
+  end
+
+  def test_sections_commands
+    section = 'Test Section'
+    doing('sections', 'add', section)
+    assert_match(/#{section}/, doing('sections', 'list'), "Sections should contain #{section}")
+
+    doing('--yes', 'sections', 'remove', section)
+    assert_no_match(/#{section}/, doing('sections', 'list'), "Sections should not contain #{section}")
+  end
+
+  private
+
+  def mktmpdir
+    tmpdir = Dir.mktmpdir
+    @tmpdirs.push(tmpdir)
+
+    tmpdir
+  end
+
+  def doing(*args)
+    doing_with_env({'DOING_CONFIG' => @config_file, 'DOING_BACKUP_DIR' => @backup_dir}, '--doing_file', @wwid_file, *args)
+  end
+end
+